+++ test/test_brain_atlas.py
@@ -1,0 +1,91 @@
+import os
+import numpy as np
+from nilearn import image
+from nilearn.input_data import NiftiMasker
+
+from photonai.base import PipelineElement
+
+from photonai_neuro import BrainMask, AtlasLibrary, BrainAtlas
+from test.test_neuro import NeuroBaseTest
+
+
+class BrainAtlasTests(NeuroBaseTest):
+
+    def test_brain_atlas(self):
+
+        brain_atlas = BrainAtlas(self.atlas_name, "vec", rois=self.roi_list)
+        new_data = brain_atlas.transform(self.X)
+        self.assertTrue(len(self.X), len(brain_atlas.rois))
+
+        brain_atlas_mean = BrainAtlas(self.atlas_name, "mean", rois='all')
+        brain_atlas_mean.transform(self.X)
+        # Todo: how to compare?
+        debug = True
+
+    def test_brain_atlas_load(self):
+
+        brain_atlas = AtlasLibrary().get_atlas(self.atlas_name)
+
+        # manually load brain atlas
+        man_map = image.load_img(os.path.join(self.atlas_folder, 'AAL_SPM12/AAL.nii.gz')).get_data()
+        self.assertTrue(np.array_equal(man_map, brain_atlas.map))
+
+    def test_custom_atlas(self):
+        custom_atlas = os.path.join(self.atlas_folder, 'AAL_SPM12/AAL.nii.gz')
+
+        atlas = PipelineElement('BrainAtlas', atlas_name=custom_atlas, extract_mode='vec', batch_size=20)
+        X_masked = atlas.transform(self.X)
+
+        with self.assertRaises(FileNotFoundError):
+            atlas = PipelineElement('BrainAtlas', atlas_name='XXXXX', extract_mode='vec', batch_size=20)
+            atlas.transform(self.X)
+
+    def test_all_atlases(self):
+        for atlas in AtlasLibrary().ATLAS_DICTIONARY.keys():
+            print("Running tests for atlas {}".format(atlas))
+            brain_atlas = PipelineElement('BrainAtlas', atlas_name=atlas, extract_mode='vec')
+            brain_atlas.transform(self.X)
+
+    def test_validity_check_roi_extraction(self):
+        for atlas in AtlasLibrary().ATLAS_DICTIONARY.keys():
+            print("Checking atlas {}".format(atlas))
+            rois = AtlasLibrary().get_atlas(atlas).roi_list[1:3]
+            rois = [roi.label for roi in rois]
+            brain_atlas = BrainAtlas(atlas_name=atlas)
+            brain_atlas.rois = rois
+            X_t = brain_atlas.transform(self.X[0:2])
+
+            "-".join(rois)
+            name = os.path.join(self.test_folder, atlas + '_' + "-".join(rois))
+            brain_atlas._validity_check_roi_extraction(X_t[0], filename=name)
+
+
+class BrainMaskTests(NeuroBaseTest):
+
+    def test_brain_masker(self):
+
+        affine, shape = BrainMask.get_format_info_from_first_image(self.X)
+        atlas_obj = AtlasLibrary().get_atlas(self.atlas_name, affine, shape)
+        roi_objects = BrainAtlas._get_rois(atlas_obj, which_rois=self.roi_list, background_id=0)
+
+        for roi in roi_objects:
+            masker = BrainMask(mask_image=roi, affine=affine, shape=shape, extract_mode="vec")
+            own_calculation = masker.transform(self.X[0])
+            nilearn_func = NiftiMasker(mask_img=roi.mask, target_affine=affine, target_shape=shape, dtype='float32')
+            nilearn_calculation = nilearn_func.fit_transform(self.X[0])
+
+            self.assertTrue(np.array_equal(own_calculation, nilearn_calculation))
+
+    def test_custom_mask(self):
+        custom_mask = os.path.join(self.atlas_folder, 'Cerebellum/P_08_Cere.nii.gz')
+        mask = PipelineElement('BrainMask', mask_image=custom_mask, extract_mode='vec', batch_size=20)
+        _ = mask.transform(self.X)
+
+        with self.assertRaises(FileNotFoundError):
+            mask = PipelineElement('BrainMask', mask_image='XXXXX', extract_mode='vec', batch_size=20)
+            mask.transform(self.X)
+
+    def test_all_masks(self):
+        for mask in AtlasLibrary().MASK_DICTIONARY.keys():
+            brain_mask = PipelineElement('BrainMask', mask_image=mask, extract_mode='vec')
+            brain_mask.transform(self.X)
