+++ test/test_neuro.py
@@ -1,0 +1,79 @@
+import os
+import numpy as np
+from nilearn import image
+from nilearn.datasets import fetch_oasis_vbm
+from sklearn.model_selection import ShuffleSplit
+
+from photonai.base import OutputSettings, Hyperpipe, PipelineElement
+from photonai.processing import ResultsHandler
+from photonai.helper.photon_base_test import PhotonBaseTest
+
+from photonai_neuro import NeuroBranch
+
+
+class NeuroBaseTest(PhotonBaseTest):
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        cls.file = __file__
+        super(NeuroBaseTest, cls).setUpClass()
+
+    @staticmethod
+    def get_data_from_oasis(n_subjects=10):
+        # GET DATA FROM OASIS
+        dataset_files = fetch_oasis_vbm(n_subjects=n_subjects)
+        age = dataset_files.ext_vars['age'].astype(float)
+        return np.array(dataset_files.gray_matter_maps), np.array(age)
+
+    def setUp(self):
+        super(NeuroBaseTest, self).setUp()
+        self.test_folder = os.path.dirname(os.path.abspath(__file__))
+        self.atlas_folder = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../photonai_neuro/atlases/')
+        self.atlas_name = "AAL"
+        self.roi_list = ["Hippocampus_R", "Hippocampus_L", "Amygdala_L", "Amygdala_R"]
+        self.X, self.y = self.get_data_from_oasis()
+
+
+class NeuroTests(NeuroBaseTest):
+
+    def test_atlas_mapper(self):
+        pass
+
+    def test_inverse_transform(self):
+        settings = OutputSettings(project_folder=self.tmp_folder_path,
+                                  overwrite_results=True)
+
+        # DESIGN YOUR PIPELINE
+        pipe = Hyperpipe('Limbic_System',
+                         optimizer='grid_search',
+                         metrics=['mean_absolute_error'],
+                         best_config_metric='mean_absolute_error',
+                         outer_cv=ShuffleSplit(n_splits=1, test_size=0.2),
+                         inner_cv=ShuffleSplit(n_splits=1, test_size=0.2),
+                         verbosity=2,
+                         cache_folder=self.cache_folder_path,
+                         eval_final_performance=True,
+                         output_settings=settings)
+
+        # PICK AN ATLAS
+        atlas = PipelineElement('BrainAtlas',
+                                rois=['Hippocampus_L', 'Amygdala_L'],
+                                atlas_name="AAL", extract_mode='vec', batch_size=20)
+
+        # EITHER ADD A NEURO BRANCH OR THE ATLAS ITSELF
+        neuro_branch = NeuroBranch('NeuroBranch')
+        neuro_branch += atlas
+        pipe += neuro_branch
+
+        pipe += PipelineElement('LinearSVR')
+
+        pipe.fit(self.X, self.y)
+
+        # GET IMPORTANCE SCORES
+        handler = ResultsHandler(pipe.results)
+        importance_scores_optimum_pipe = handler.results.best_config_feature_importances
+
+        manual_img, _, _ = pipe.optimum_pipe.inverse_transform(importance_scores_optimum_pipe, None)
+        img = image.load_img(os.path.join(self.tmp_folder_path,
+                                          'Limbic_System_results/optimum_pipe_feature_importances_backmapped.nii.gz'))
+        self.assertTrue(np.array_equal(manual_img.get_data(), img.get_data()))
